+++ src/Termbox/Cursor.hs
@@ -1,17 +1,12 @@
 module Termbox.Cursor
-  ( hideCursor,
-    setCursor,
+  ( Cursor (..),
   )
 where
 
 import Termbox.Internal
 
--- | Set the cursor coordinates (column, then row).
-setCursor :: Int -> Int -> IO ()
-setCursor =
-  tb_set_cursor
-
--- | Hide the cursor.
-hideCursor :: IO ()
-hideCursor =
-  tb_set_cursor tB_HIDE_CURSOR tB_HIDE_CURSOR
+-- | A cursor.
+data Cursor
+  = -- | Column, then row
+    Cursor !Int !Int
+  | NoCursor