+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to the [Haskell Package Versioning Policy](https://pvp.
 
 ### Added
 - Add `defaultInputMode`, `defaultOutputMode`
+- Add `Cells`, `Cursor`, `render`
 - Export `Termbox.Internal` module that (roughly) contains 1:1 C bindings
 
 ### Changed
@@ -16,9 +17,10 @@ and this project adheres to the [Haskell Package Versioning Policy](https://pvp.
 - Add `InputMode` and `OutputMode` arguments to `run`
 - Reset output mode to "normal" on shutdown to work around a small bug in termbox.c that retains the output mode across
   separate invocations of init/shutdown
+- Change type of `set` to construct a `Cells` rather than an `IO ()`
 
 ### Removed
-- Remove `getInputMode`, `setInputMode`, `getOutputMode`, `setOutputMode`
+- Remove `getInputMode`, `setInputMode`, `getOutputMode`, `setOutputMode`, `getCursor`, `hideCursor`, `clear`, `flush`
 - Remove build dependency on `c2hs`
 
 ## [0.2.0.1] - 2020-06-27